StarBattle: __add_line_constraints
Use a method to calculate the contraints for the number of stars in
a row or column. When calculating the constraints for a cell, we
can calculate how many stars can there be at most to its left, and
above it, and for cells close the right or bottom edge, the minimum
number of stars which should be to the left of it, or above it.